Move POS onboarding shown/dismissed event tracking to `TotalsViewModel`.
