+++ WooCommerce/Classes/POS/Presentation/Payments Onboarding/PointOfSaleCardPresentPaymentOnboardingView.swift
@@ -14,9 +14,6 @@ struct PointOfSaleCardPresentPaymentOnboardingView: View {
                              accessibilityLabel: Localization.cancelOnboarding)
         .safariSheet(url: $viewModel.onboardingURL)
         .posModalSizing()
-        .onAppear {
-            viewModel.trackOnboardingShown()
-        }
     }
 }
