+++ WooCommerce/WooCommerceTests/POS/ViewModels/TotalsViewModelTests.swift
@@ -294,6 +294,46 @@ final class TotalsViewModelTests: XCTestCase {
         // Then
         XCTAssertEqual(sut.cardPresentPaymentOnboardingViewModel?.state, .pluginNotActivated(plugin: .stripe))
     }
+
+    // MARK: Analytics
+
+    func test_paymentsOnboardingDismissed_event_is_tracked_with_state_when_cancelOnboarding_is_invoked() throws {
+        // Given
+        let analyticsProvider = MockAnalyticsProvider()
+        let analytics = WooAnalytics(analyticsProvider: analyticsProvider)
+        let sut = TotalsViewModel(orderService: orderService,
+                                  cardPresentPaymentService: cardPresentPaymentService,
+                                  currencyFormatter: .init(currencySettings: .init()),
+                                  paymentState: .acceptingCard,
+                                  analytics: analytics)
+        let onboardingViewModel = CardPresentPaymentsOnboardingViewModel(fixedState: .noConnectionError)
+        cardPresentPaymentService.paymentEvent = .showOnboarding(onboardingViewModel: onboardingViewModel, onCancel: {})
+
+        // When
+        sut.cancelOnboarding()
+
+        // Then
+        XCTAssertNotNil(analyticsProvider.receivedEvents.first(where: { $0 == "pos_payments_onboarding_dismissed" }))
+        let eventProperties = try XCTUnwrap(analyticsProvider.receivedProperties.first(where: { $0.keys.contains("onboarding_state") }))
+        XCTAssertEqual(eventProperties["onboarding_state"] as? String, "no_connection_error")
+    }
+
+    func test_pointOfSalePaymentsOnboardingShown_event_is_tracked_when_trackOnboardingShown_is_invoked() throws {
+        // Given
+        let analyticsProvider = MockAnalyticsProvider()
+        let analytics = WooAnalytics(analyticsProvider: analyticsProvider)
+        let sut = TotalsViewModel(orderService: orderService,
+                                  cardPresentPaymentService: cardPresentPaymentService,
+                                  currencyFormatter: .init(currencySettings: .init()),
+                                  paymentState: .acceptingCard,
+                                  analytics: analytics)
+
+        // When
+        sut.trackOnboardingShown()
+
+        // Then
+        XCTAssertNotNil(analyticsProvider.receivedEvents.first(where: { $0 == "pos_payments_onboarding_shown" }))
+    }
 }
 
 private extension TotalsViewModelTests {
